EnqueuedResourceOffloading: remove unused properties
This commit removes the `EnqueuedResourceOffloadingSniff::$false_tokens` and the `EnqueuedResourceOffloadingSniff::$safe_tokens` properties as they are not used by the sniff. I believe they were introduced accidentally due to a copy and paste error when creating the sniff.

Since those properties were removed, it is not necessary anymore to overload the `register()` method to change the value of those properties, so this method was removed as well.

This change should not affect the functionality of the sniff.